api/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-learning-api"
version = "0.1.0"
description = "FastAPI backend for the RAG learning platform"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.12,<1.0.0",
  "uvicorn[standard]>=0.34.2,<1.0.0",
  "sqlalchemy>=2.0.40,<3.0.0",
  "asyncpg>=0.30.0,<1.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "httpx>=0.28.1,<1.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "pgvector>=0.3.6,<1.0.0",
  "google-genai>=0.5.0,<1.0.0",
  "pandas>=2.0.0,<3.0.0",
  "duckdb>=1.0.0,<2.0.0",
  "openpyxl>=3.0.0,<4.0.0",
  "xlrd>=2.0.1,<3.0.0",
  "langchain-text-splitters>=0.3.6,<1.0.0",
  "pymupdf>=1.25.3,<2.0.0",
  "pypdf>=5.0.0,<6.0.0",
  "python-docx>=1.1.2,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5,<9.0.0",
  "pytest-asyncio>=0.26.0,<1.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["api*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-p no:cacheprovider"
